+++ src/main/java/com/bookstore/project/repository/BookRepository.java
@@ -10,6 +10,6 @@ import java.util.List;
 
 public interface BookRepository extends JpaRepository<Book, Integer> {
     List<Book> findByGenreId(int genreId);
-    List<Book> findByUserId(int userId);
+    List<Book> findByUserId(long userId);
 }
